Minor repairs (#11)
Fixes the odd use of 128-bit long ints in the opencl code (some compilers/cards did not like).
Perhaps a more universal method for getting ray (distributed processing) to initialize on local machine, even when on a VPN. The older method was having some issues with some firewalls.
Signed-off by: David Rowenhorst <david.rowenhorst@nrl.navy.mil>

+++ pyebsdindex/clkernels.cl
@@ -155,8 +155,8 @@ __kernel void convolution3d2d( __global const float16 *in, __constant float *ker
   const unsigned long int nImChunk = get_global_size(2);
   //long int indxIm; 
   //long int indxK;
-  long unsigned long yIndx; 
-  long unsigned long yKIndx; 
+  unsigned long yIndx; 
+  unsigned long yKIndx; 
   float16 pxVal;
   float16 sum = (float16) (0.0f); // initialize convolution sum
   float kVal;
